requirements.txt
requests==2.30.0
loguru==0.7.0
python-dotenv==1.0.0
python-telegram-bot==20.3
